resolve pr issues
