+++ src/api/types.ts
@@ -6,3 +6,18 @@ export type FusionApiConfig = {
     network: NetworkEnum
     httpProvider?: HttpProviderConnector
 }
+
+export type PaginationMeta = {
+    totalItems: number
+    itemsPerPage: number
+    totalPages: number
+    currentPage: number
+}
+
+export type PaginationOutput<
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    T extends Record<string, any> = Record<string, any>
+> = {
+    meta: PaginationMeta
+    items: T[]
+}
