--- src/types.ts
@@ -1,14 +1,0 @@
-export type PaginationMeta = {
-    totalItems: number
-    itemsPerPage: number
-    totalPages: number
-    currentPage: number
-}
-
-export type PaginationOutput<
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    T extends Record<string, any> = Record<string, any>
-> = {
-    meta: PaginationMeta
-    items: T[]
-}
